requirements.txt
torch>=2.2.0
torchaudio>=2.2.0
demucs>=4.0.0
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
requests>=2.31.0
sseclient-py>=1.7.0
numpy
werkzeug>=2.3.7
diffq
psutil>=5.9.0 